Add a new method to send a HTTP POST request to our servers.
This request will contain all information about any dialogue line, that doesn't have any audio file associated.
I find it a bit hard to move around in the code as I'm not the one who wrote it, so I left a comment for you with clear instructions on what to do to make this work.
The server-side part is fully functional now, it's just missing some tools to work with the saved reports.